+++ src/main/kotlin/no/nav/bidrag/beregn/nettobarnetilsyn/bo/periode.kt
@@ -2,13 +2,14 @@ package no.nav.bidrag.beregn.nettobarnetilsyn.bo
 
 import no.nav.bidrag.beregn.felles.bo.Periode
 import no.nav.bidrag.beregn.felles.bo.PeriodisertGrunnlag
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class FaktiskUtgiftPeriode(
     val faktiskUtgiftSoknadsbarnPersonId: Int,
     val faktiskUtgiftDatoFraTil: Periode,
     val faktiskUtgiftSoknadsbarnFodselsdato: LocalDate,
-    val faktiskUtgiftBelop: Double) : PeriodisertGrunnlag {
+    val faktiskUtgiftBelop: BigDecimal) : PeriodisertGrunnlag {
   constructor(faktiskUtgiftPeriode: FaktiskUtgiftPeriode)
       : this(faktiskUtgiftPeriode.faktiskUtgiftSoknadsbarnPersonId,
       faktiskUtgiftPeriode.faktiskUtgiftDatoFraTil.justerDatoer(),
