+++ src/test/java/no/nav/bidrag/beregn/nettobarnetilsyn/NettoBarnetilsynCoreTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -108,7 +109,7 @@ public class NettoBarnetilsynCoreTest {
 
     var faktiskUtgiftPeriode = new FaktiskUtgiftPeriodeCore(
         new PeriodeCore(LocalDate.parse("2017-01-01"), null), LocalDate.parse("2010-01-01"),
-        1, 2);
+        1, BigDecimal.valueOf(2));
     var faktiskUtgiftPeriodeListe = new ArrayList<FaktiskUtgiftPeriodeCore>();
     faktiskUtgiftPeriodeListe.add(faktiskUtgiftPeriode);
 
@@ -127,9 +128,9 @@ public class NettoBarnetilsynCoreTest {
 
     periodeResultatListe.add(new ResultatPeriode(
         new Periode(LocalDate.parse("2017-01-01"), LocalDate.parse("2018-01-01")),
-        Arrays.asList(new ResultatBeregning(1, 1)),
+        Arrays.asList(new ResultatBeregning(1, BigDecimal.valueOf(1))),
         new GrunnlagBeregningPeriodisert(
-            Arrays.asList(new FaktiskUtgift(1, 10,3)),
+            Arrays.asList(new FaktiskUtgift(1, 10, BigDecimal.valueOf(3))),
             Arrays.asList(new Sjablon(SjablonTallNavn.SKATTESATS_ALMINNELIG_INNTEKT_PROSENT.getNavn(), emptyList(),
                 Arrays.asList(new SjablonInnhold(SjablonInnholdNavn.SJABLON_VERDI.getNavn(), 22d)))))));
